requirements.txt
Cerberus==1.3.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Injector==0.12.0
idna==2.8
inflection==0.3.1
injector==0.16.2
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.7
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
swagger-ui-bundle==0.0.3
typing==3.6.6
urllib3==1.25.3
Werkzeug==0.15.4
gunicorn==19.9.0
python-dateutil==2.7.2
beautifulsoup4==4.4.1
pdfkit==0.6.1
pyiec==0.1.2

